+++ test/raxx/http1_test.exs
@@ -1,0 +1,4 @@
+defmodule Raxx.HTTP1Test do
+  use ExUnit.Case
+  doctest Raxx.HTTP1
+end