+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.15.7](https://github.com/CrowdHailer/raxx/tree/0.15.7) - 2018-08-18
+
+### Added
+
+- `Raxx.HTTP1` Tools for parsing and serializing to HTTP1 format.
+
 ## [0.15.6](https://github.com/CrowdHailer/raxx/tree/0.15.6) - 2018-08-09
 
 ### Fixed
 
-- `Raxx.request` sets the raw_path as `/` when no path component given. 
+- `Raxx.request` sets the raw_path as `/` when no path component given.
 
 ## [0.15.5](https://github.com/CrowdHailer/raxx/tree/0.15.5) - 2018-08-04